Fix issues with the dropdowns on mobile
v-click-outside uses the "touchstart" event to detect clicks/touches outside of the specified element on mobile. Since touchstart occurs at the beginning of the touch and "click" occurs at the end, the menu was being hidden before the click event could go through.
This actually may have still worked if I had used "touchend" but it looks better for the menu to stay open until the touch ends, so @touchstart.stop prevents the event from propagating up to v-touch-outside.